zerobus-ingest/pyproject.toml
[project]
name = "zerobus-ingest"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bufbuild-protovalidate-protocolbuffers-python==33.5.0.1.20260209202127+80ab13bee0bf",
    "databricks-sdk>=0.88.0",
    "databricks-connect>=18.0.0",
    "databricks-zerobus-ingest-sdk>=0.2.0",
    "delta-spark==4.0.1",
    "grpcio>=1.78.0",
    "grpcio-status>=1.78.0",
    "protobuf>=6.33.0",
    "protovalidate>=1.1.1",
    "python-dotenv>=1.2.1",
]

[tool.uv]
extra-index-url = ["https://buf.build/gen/python"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 105
src = ["src", "tests", "main.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.1",
]

[project.urls]
Repository = "https://github.com/newfront/learning-spark-sdp"
Releases = "https://github.com/newfront/learning-spark-sdp/releases"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src", "gen/python"]
exclude = ["*.__pycache__", "orders.v1.__pycache__"]

[tool.setuptools.package-dir]
"" = "src"
"orders" = "gen/python/orders"

# Include descriptor.bin and .pyi in the orders.v1 package (wheel + sdist)
[tool.setuptools.package-data]
"orders.v1" = ["*.bin", "*.pyi"]
